Clean up Proxy-based stuff a little bit.
Removed a bit of repetition in Proxy-based code.
We can use type inference to our advantage and
get rid of type annotations. Also, we can define
implicit conversions in Base trait as they simply
forward to proxyOps method that is defined in a
base trait too.

Sadly enough, we canot do any better which is
confirmed by @adriaanm.

Signed-off-by: Grzegorz Kossakowski <grzegorz.kossakowski@gmail.com>