Implements the DOM API in Scala, as needed for running the Koch example.
The implementation is tailored to running the Koch example, which
makes use of one Canvas and its 2D Context. It assumes that:
- there is only one final call to stroke (so that the lines can be
  drawn online);
- closePath is a formality (i.e. the lines already draw a close path);
- there is only one canvas.

The implementation uses Java's java.awt.Graphics2D API. It is pretty
simple, because both Java's API and the JS Canvas API have a notion of
transforms. It's easy to emulate the JS Canvas save() - restore()
stack by an explicit stack of transforms. Another difference is that
the JS Canvas API has a current point state (drawing a line uses this
state as a starting point and updates it); we emulate this current
point state with a field in Scala: the only tricky bit is to make sure
we draw the point correctly with respect to the current transformation
(this is done by also storing the transform that was current when the
point was updated and ensuring that it gets draw with respect to that
transform).