Basic reified traits, generating working JS code.
The trait is reified into an IR-node by explicit registration in the
user code. All methods are invoked, and reified during registration at
staging time, and the IR-node for the trait is just a bundle of those
reified methods. Registering a trait returns a factory for a Rep of
this trait.

Not yet implemented:
- no sanity checks
  (e.g. all method parameter types should be Reps, method return type
  should be Rep, trait to reify should be instantiable (no abstract
  methods or vars), no vals are allowed (because the proxy doesn't
  handle them correctly: it calls them with extra empty parentheses))
- inheritance
- trivial in-Scala embedding